Parse yield with no argument in an AssignmentExpression.
Summary:
It is possible for `YieldExpression` to be used as a part of a greater
AssignmentExpression without itself having an AssignmentExpression as
an argument.
For example:
```
x ? yield : y
```
is valid JS in a generator function.

Previously, we were only checking for `eatSemi` before attempting to
parse an argument to `yield`, while we should have been detecting whether
the next token could begin an AssignmentExpression.

As a simpler fix, we check whether the next token can *end* the AssignmentExpression
which the YieldExpression matches by adding a new function which checks
for all possible tokens which can end an AssignmentExpression.

Make this easier to maintain by asserting every time we parse an
AssignmentExpression that the token is one of the tokens in that list.

Reviewed By: tmikov

Differential Revision: D16774445

fbshipit-source-id: b3ee48672fa3c91cd80270d6ca0b97f686bf4de1