Turn AssignmentExpression end assert into a real check.
Summary:
The assert at the end of parseAssignmentExpression to make sure
that the `checkEndAssignmentExpression` function was kept up to date
would result in crashes if invalid JS was provided to the parser.

To avoid this, we turn that assertion into an actual check with an error,
so that we will report the correct failure.

Now, if the `checkEndAssignmentExpression` function comes out of date,
the parser will report an error instead of crashing but the signal will
still indicate to update `checkEndAssignmentExpression`.

Reviewed By: tmikov

Differential Revision: D17371201

fbshipit-source-id: 473a95b29700349b31c7bab99daf1311de3bd871