Only store partial ID in DictPropertyMap's hash table
Summary:
Halve the size of each entry in the hash table of DictPropertyMap by instead of storing a SymbolID + a (32-bit) descriptor index, only storing part of the ID (for filtering) and using fewer bits for the descriptor index (as allowed by the max capacity of DPM). The partial symbol ID allows for a first fast filtering.

Introduce an abstraction for HashPair that also checks the allowed state transitions of an entry (empty -> valid -> deleted -> valid -> ...)

Reviewed By: tmikov

Differential Revision: D17327870

fbshipit-source-id: aeb3667a4dbfe23020046e6849aca935a3a3a86f