Fix CreateArguments lowering in generator functions.
Summary:
`GeneratorInnerFunction`s have their `CreateArgumentsInst` in the second
code block, because the first code block is used for saving initial
state and argument creation shouldn't take place until the user calls
`next` for the first time on the generator.

The `CreateArgumentsInst` lowering was assuming that the instruction
was always in the first block, missing it, and failing to lower it.

As such, we add a branch and move to the second code block in
`GeneratorInnerFunction`s only.

Reviewed By: tmikov

Differential Revision: D16226681

fbshipit-source-id: 5ee5652bd8fe0fbffacc31460c6638059327b4a1