Measure private dirty memory when printing GC stats
Summary:
As a complement to our RSS metrics, which are somewhat noisy, add a "private dirty" metric obtained by parsing /proc/self/smaps. It's basically the subset of RSS that cannot be paged out (in particular, it excludes the bytecode file). This seems to reduce the noise by ~100x on our large RN benchmarks.

Not yet implemented for Windows.

Reviewed By: tmikov

Differential Revision: D17504910

fbshipit-source-id: d651a5ce9312bc307e53b59f7b0b458622a2c2df