Add function offsets to SourceMapGenerator
Summary: Add support for keeping track of function offsets in SourceMapGenerator. Update chained calls that copy `functionOffsets` to move ownership of the vector to the callee, in order to store the offsets at `SourceMapGenerator` (without expensive copying) as it is a stateful SourceMap generator class. Output`functionsOffsets` under `x_hermes_function_offsets` in `SourceMapGenerator::outputAsJSON`.

Reviewed By: motiz88

Differential Revision: D17474360

fbshipit-source-id: f9c4b7cdef185a173e23b7f3c02ce6d9c8992bc7