Make occupancyTarget configurable
Summary:
A major parameter of the heap growth policy is the occupancy target - the target fraction of live data in the heap. A higher value typically means more conservative growth and thus also more frequent collections.

Expose this as a configurable parameter (both in RuntimeConfig API and in the command-line runners).

Reviewed By: davedets

Differential Revision: D17604669

fbshipit-source-id: 1160ff83af1b8d188a799ca943a79f64dba60eb4