Add function to canonicalize characters and code point sets
Summary:
Canonicalization is an ES operation which maps a character to uppercase
(or case folds if Unicode is set).

Prior to this fix, canonicalization was implemented by calling into the
platform's native Unicode implementation. However Android does not have
a native case folding operation, so we will need to do something
different there.

For simplicity, emit a table for canonicalization and add functions to
canonicalize and "precanonicalize" characters.

Reviewed By: avp

Differential Revision: D17413827

fbshipit-source-id: a4adc2dc721eff79f06df935cac35cf12eda3eda