Fix AnchoredAtStart constraints on lookbehinds
Summary:
The Hermes regex engine has a notion of "constraints" which allow for
quickly rejecting matches. One such constraint is that a regex may only
match at the start, due to a ^ anchor. For example:

    /^abc/

can only match abc at the beginning of the string, so there is no need to
search from the second character onward.

Lookahead assertions may declare this constraint, for example:

    /(?=^ab)c/

matches the same as above.

But lookbehind assertions do not. Because they walk the input string
backwards, it's possible to reach the beginning even if we don't start
there. For example:

    /(?<=^ab)c/

matches the string "abc" but only starting at the c. The anchor does not
require that the match be at the start.

This fix ensures lookbehind assertions do not declare that they are
anchored at the start.

Reviewed By: willholen

Differential Revision: D17608484

fbshipit-source-id: d51c6dece7548997564601a32fb42e018a22aaf3