Eliminate prevCharAvailable and move satisfiesConstraints to Cursor
Summary:
Previously a regex client could supply the `prevCharAvailable` flag, which
indicated that the character at index -1 in the index string was valid.
This supported word boundary instructions.

Now that we require that clients supply the entire string to the engine,
this flag should never be set, so eliminate it.

This flag also supported an optimization: it tracked whether a regex could
only match at the beginning of a string. Now the knowledge of whether we
are at the beginning of a string resides in Cursor, so migrate the
constraint satisfaction logic to Cursor.

Reviewed By: avp

Differential Revision: D17608479

fbshipit-source-id: 57d7e221ae94bc30533367078ff31220141a2fee