Fix calculation of peakLiveAfterGC
Summary:
D17833692 introduced tracking the maximum of used size before/after after GC, and exposed this data as "peakAllocatedBytes" and "peakLiveAfterGC".

The latter was incorrectly calculated for the generational GC. Measuring "live" data only makes sense after a *full* GC, but the code was looking at GCBase::cumStats_ which contains an indiscriminate mix of young and full collections. (This is in contrast with "peakAllocatedBytes", which can legitimately occur before any kind of collection, hence GCBase::cumStats_ is correct there.)

Fix the issue by refactoring the code to allow overriding peakLiveAfterGC for GenGC.

Reviewed By: davedets

Differential Revision: D17847598

fbshipit-source-id: 605a40050ac3801b8fd36c2b8e1817f5f830c5df