Add support for spread args.
Summary:
Allow the spread element in arguments using a simple naive implementation.
We add a `HermesInternal.apply` function, spread the arguments into
an array using the spread array mechanisms, and then use HermesInternal.apply
to actually run the call.

Note that because spread args can be used in constructors, we need to
convey the use of `new` to `HermesInternal.apply` with one of the arguments.

Reviewed By: tmikov

Differential Revision: D17334064

fbshipit-source-id: f358475122bac76cc1a6533ac8ccd6d62851d312